LayoutTests/platform/glib/imported/w3c/web-platform-tests/fetch/cross-origin-resource-policy/scheme-restriction.any.worker-expected.txt

CONSOLE MESSAGE: Cancelled load to https://www1.web-platform.test:9443/fetch/cross-origin-resource-policy/resources/hello.py?corp=same-site because it violates the resource's Cross-Origin-Resource-Policy response header.

PASS Cross-Origin-Resource-Policy: same-site blocks retrieving HTTPS from HTTP
